requirements.txt
amqp==5.3.1
annotated-types==0.7.0
awscli-local==0.22.0
billiard==4.2.1
blinker==1.9.0
boto3==1.35.60
botocore==1.35.60
celery==5.4.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
docx2txt==0.8
easyocr==1.7.2
et_xmlfile==2.0.0
filelock==3.16.1
Flask==3.1.0
fsspec==2024.10.0
huggingface-hub==0.26.2
idna==3.10
imageio==2.36.0
itsdangerous==2.2.0
Jinja2==3.1.4
jmespath==1.0.1
kombu==5.4.2
lazy_loader==0.4
localstack-client==2.7
lxml==5.3.0
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.4.2
ninja==1.11.1.1
numpy==2.1.3
opencv-python-headless==4.10.0.84
openpyxl==3.1.5
packaging==24.2
pdf2image==1.17.0
pillow==11.0.0
prompt_toolkit==3.0.48
psycopg2-binary==2.9.10
pyclipper==1.3.0.post6
pydantic==2.9.2
pydantic_core==2.23.4
pytesseract==0.3.13
python-bidi==0.6.3
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.1
PyYAML==6.0.2
redis==5.2.0
regex==2024.11.6
requests==2.32.3
s3transfer==0.10.3
safetensors==0.4.5
scikit-image==0.24.0
scipy==1.14.1
setuptools==69.1.0
shapely==2.0.6
six==1.16.0
SQLAlchemy==2.0.36
sympy==1.13.1
tifffile==2024.9.20
tokenizers==0.20.3
torch==2.5.1
torchvision==0.20.1
tqdm==4.67.0
transformers==4.46.2
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
vine==5.1.0
wcwidth==0.2.13
Werkzeug==3.1.3
wheel==0.42.0
